+++ src/CHANGELOG.md
@@ -7,3 +7,7 @@ A list of important changes that may break things
 3. `PUT` `/lock/<string:customer_id>` now returns `204` instead of `200`
 3. `DELETE` `/lock/<string:customer_id>` now returns `204` instead of `200`
 4. `PUT` `/subscription` now returns `204` instead of `200`
+5. `POST` `/task/assign` now returns `204` instead of `200`
+6. `POST` `/task/cancel` now returns `204` instead of `200`
+7. `POST` `/task/delay` now returns `204` instead of `200`
+8. `GET` `/task/delay/{id}` returns `delayed_by` as str not User